Maintain original version of python demo
